backend/requirements.txt
django
djongo
djangorestframework
dnspython
django-cors-headers
djangorestframework-simplejwt
scikit-learn
